pyproject.toml
[project]
name = "mcp-cabs-search-book"
version = "0.1.0"
description = "Tools for search and book cabs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.0.0b1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "googlemaps>=4.10.0",
    "httpx>=0.27.0",
]
[tool.hatch.build.targets.wheel]
packages = ["src/mcp-cab-server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
